+++ src/sources/external_input_video.ml
@@ -132,11 +132,11 @@ let () =
         Lang.bool_t,
         Some (Lang.bool false),
         Some "Restart process when exited with error." );
-      ("", Lang.string_t, None, Some "Command to execute.");
+      ("", Lang.getter_t Lang.string_t, None, Some "Command to execute.");
     ]
     ~return_t
     (fun p ->
-      let command = Lang.to_string (List.assoc "" p) in
+      let command = Lang.to_string_getter (List.assoc "" p) in
       let video_format = ref None in
       let width = ref None in
       let height = ref None in
@@ -290,11 +290,11 @@ let () =
         Lang.bool_t,
         Some (Lang.bool false),
         Some "Restart process when exited with error." );
-      ("", Lang.string_t, None, Some "Command to execute.");
+      ("", Lang.getter_t Lang.string_t, None, Some "Command to execute.");
     ]
     ~return_t
     (fun p ->
-      let command = Lang.to_string (List.assoc "" p) in
+      let command = Lang.to_string_getter (List.assoc "" p) in
       let width = Lazy.force Frame.video_width in
       let height = Lazy.force Frame.video_height in
       let buflen = width * height * 3 in