+++ CHANGES.md
@@ -12,6 +12,7 @@ Changed:
 * Removed `disconnect` method on `input.harbor`. This method was doing the same as
   the `stop` method. Added `shutdown` method to properly shutdown the
   source even when not connected to an output.
+* Made process a string getter in `input.external.{rawaudio,rawvideo}` (#2877)
 
 Fixed:
 * Fixed parameter type for `stats_interval` in SRT I/O.